Make distcheck target work for unprivileged users
The integration with SANE and/or udev requires write access to system
locations that are beyond the control of Utsushi.  A distcheck target
runs various install related targets that attempt to make changes in
these locations.  While a $(DESTDIR) version of these targets should
(and does) work fine, the regular versions do not in the absence of
administrative privileges.  Disabling this integration for distcheck
targets is preferred over requiring administrative privileges.