Download Boost source archive from upstream site
This addresses a documented "bug" in tools/include-boost-src and makes
it generally easier for developers to include and/or upgrade the Boost
sources.  The preferred Boost version has been changed to 1.46.1 since
that was the first version recent enough that was distributed as part
of an Ubuntu release after 10.04 (for which the upstream binaries were
built).