Put upstream downloads in a fixed location
Although trying to do this using a location relative to the top source
directory is understandable, it creates a download directory one level
above that directory and triggers a second download when doing a make
distcheck.  This changeset adds a configuration option (for developer
and maintainer types) that sets an absolute directory as the download
location.  The default is upstream/downloads in the top level source
directory.  This no longer creates a second location elsewhere, nor
does it trigger a second download during a distcheck.