Add a default device query to utsushi::monitor
The utsushi::monitor singleton maintains information on all scanner
devices, including those that are not supported yet or for which a
driver has not been installed.  Hence, the size() member function
cannot be used to determine whether an application can scan.  The
new query returns a non-empty string if it can.

This fixes unit test failures when tests are run with unsupported
devices attached.