Add support for (almost) all hardware options
This does not take into account whether options are actually supported
by the device yet.  Nor does it account for documented restrictions on
the option values.

Support for custom gamma profiles, color matrices and dither patterns
is still missing, as is an option to control the main lamp's lighting
mode (whatever that may be).